v1.1.0

+++ src/main/kotlin/us/timinc/mc/cobblemon/granularshinies/config/MainConfig.kt
@@ -2,5 +2,5 @@ package us.timinc.mc.cobblemon.granularshinies.config
 
 class MainConfig {
     val debug: Boolean = false
-    val overrides: Map<String, Double> = mutableMapOf()
+    val overrides: Map<String, Float> = mutableMapOf()
 }
